refactoring: resolvers, typeDefs를 ./model/...로 이동
squahs there
