+++ backend/node/graphQL/app.js
@@ -3,8 +3,8 @@ import cors from "cors";
 import {GraphQLServer} from "graphql-yoga";
 import morgan from "morgan";
 import cookieParser from "cookie-parser";
-import resolvers from "./resolvers.js";
-import typeDefs from "./typeDefs.js";
+import resolvers from "./model/resolvers.js";
+import typeDefs from "./model/typeDefs.js";
 import config from "./config.js";
 import logger from "./logger.js";
 import authenticate from "./middlewares/authenticate.js";
