+++ backend/node/spec/graphql/emoji.test.js
@@ -3,8 +3,8 @@ import {after, before, beforeEach, describe, it} from "mocha";
 import gql from "graphql-tag";
 import EasyGraphQLTester from "easygraphql-tester";
 import emojiResolvers from "../../graphQL/model/emoji/emoji.resolver.js";
-import typeDefs from "../../graphQL/typeDefs.js";
-import resolvers from "../../graphQL/resolvers.js";
+import typeDefs from "../../graphQL/model/typeDefs.js";
+import resolvers from "../../graphQL/model/resolvers.js";
 import SequelizeTestHelper from "../testHelper/SequelizeTestHelper.js";
 import {createEmoji} from "../../DB/queries/emoji.js";
 import models from "../../DB/models";
