+++ backend/node/spec/graphql/guest.test.js
@@ -4,8 +4,8 @@ import EasyGraphQLTester from "easygraphql-tester";
 import {after, before, beforeEach, describe, it} from "mocha";
 import guestResolvers from "../../graphQL/model/guest/guest.resolver.js";
 import SequelizeTestHelper from "../testHelper/SequelizeTestHelper.js";
-import typeDefs from "../../graphQL/typeDefs.js";
-import resolvers from "../../graphQL/resolvers.js";
+import typeDefs from "../../graphQL/model/typeDefs.js";
+import resolvers from "../../graphQL/model/resolvers.js";
 import models from "../../DB/models";
 import {createGuest} from "../../DB/queries/guest.js";
 import {findOrCreateEvent} from "../../DB/queries/event.js";
