+++ backend/node/spec/graphql/like.test.js
@@ -2,8 +2,8 @@ import assert from "assert";
 import gql from "graphql-tag";
 import {after, before, beforeEach, describe, it} from "mocha";
 import EasyGraphQLTester from "easygraphql-tester";
-import typeDefs from "../../graphQL/typeDefs.js";
-import resolvers from "../../graphQL/resolvers.js";
+import typeDefs from "../../graphQL/model/typeDefs.js";
+import resolvers from "../../graphQL/model/resolvers.js";
 import SequelizeTestHelper from "../testHelper/SequelizeTestHelper.js";
 import models from "../../DB/models";
 import likeResolvers from "../../graphQL/model/like/like.resolver.js";
