+++ backend/node/spec/graphql/question.test.js
@@ -2,8 +2,8 @@ import assert from "assert";
 import gql from "graphql-tag";
 import EasyGraphQLTester from "easygraphql-tester";
 import {after, before, beforeEach, describe, it} from "mocha";
-import typeDefs from "../../graphQL/typeDefs.js";
-import resolvers from "../../graphQL/resolvers.js";
+import typeDefs from "../../graphQL/model/typeDefs.js";
+import resolvers from "../../graphQL/model/resolvers.js";
 import SequelizeTestHelper from "../testHelper/SequelizeTestHelper.js";
 import models from "../../DB/models";
 import {createGuest} from "../../DB/queries/guest.js";
